+++ nodes/__init__.py
@@ -4,6 +4,7 @@ from .aio_generate import AIOImageGenerate
 from .flux2_klein_settings import AIOFlux2Klein9BSettings
 from .ideogram4_prompt_builder import AIOIdeogram4PromptBuilder
 from .ideogram4_settings import AIOIdeogram4Settings
+from .info import AIOInpaintInfo, AIOModelInfo, AIOPIDInfo
 from .inpaint import AIOInpaint
 from .krea2_settings import AIOKrea2Settings
 from .lora_configuration import AIOLoraConfiguration
@@ -14,9 +15,12 @@ __all__ = [
     "AIOFlux2Klein9BSettings",
     "AIOIdeogram4PromptBuilder",
     "AIOIdeogram4Settings",
+    "AIOInpaintInfo",
     "AIOImageGenerate",
     "AIOInpaint",
     "AIOKrea2Settings",
+    "AIOModelInfo",
+    "AIOPIDInfo",
     "AIOLoraConfiguration",
     "AIOLoadPipelineModels",
     "AIOZImageTurboSettings",